#30 @tayanefernandes @josigama @roberto @otaviocarvalho @that-luiz - move connection db for app.js and add tests
