+++ tests/app_spec.js
@@ -1,0 +1,19 @@
+var app = require('../src/app');
+
+var restify = require('restify');
+
+describe('App', function () {
+  var server;
+
+  beforeEach(function () {
+    server = createSpyObj('server', ['get', 'listen']);
+    spyOn(restify, 'createServer').andReturn(server);
+  });
+
+  describe('start', function () {
+    it('starts the server', function () {
+      app.start();
+  //    expect(server.listen).toHaveBeenCalled();
+    });
+  });
+});
